ctreate the widgate folder and start to add the compoments to it
