+++ frontend/src/widgets/auth-widgest/index.ts
@@ -1,0 +1,4 @@
+export * from "./forget-password/ui/forget-password";
+export * from "./login/ui/login-form";
+export * from "./register/ui/register-form";
+export * from "./email/ui/enter-email";
