+++ frontend/src/widgets/header/ui/search-bar.tsx
