+++ frontend/src/widgets/home-widgets/catagory/ui/catagory-card.tsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useRef, useState } from "react";
+"use client ";
+import { useRef, useState, useEffect } from "react";
 
 function useReveal<T extends HTMLElement>(options?: IntersectionObserverInit) {
     const ref = useRef<T | null>(null);
@@ -22,8 +23,7 @@ function useReveal<T extends HTMLElement>(options?: IntersectionObserverInit) {
 
     return { ref, visible };
 }
-
-type Card = {
+export type Card = {
     title: string;
     emoji: string;
     colorClass: string; // bg-* token you already use
@@ -31,50 +31,6 @@ type Card = {
     description: string;
 };
 
-const CARDS: Card[] = [
-    {
-        title: "Leafy Greens",
-        emoji: "🥬",
-        colorClass: "bg-primary",
-        buttonClass: "btn-primary",
-        description: "Fresh spinach, lettuce, kale and more",
-    },
-    {
-        title: "Root Vegetables",
-        emoji: "🥕",
-        colorClass: "bg-secondary",
-        buttonClass: "btn-secondary",
-        description: "Carrots, potatoes, onions and more",
-    },
-    {
-        title: "Fresh Produce",
-        emoji: "🥒",
-        colorClass: "bg-accent",
-        buttonClass: "btn-outline",
-        description: "Cucumbers, tomatoes, peppers and more",
-    },
-    {
-        title: "Fruits",
-        emoji: "🍎",
-        colorClass: "bg-primary",
-        buttonClass: "btn-primary",
-        description: "Apples, bananas, berries and more",
-    },
-    {
-        title: "Herbs",
-        emoji: "🌿",
-        colorClass: "bg-secondary",
-        buttonClass: "btn-secondary",
-        description: "Mint, basil, coriander and more",
-    },
-    {
-        title: "Organic Boxes",
-        emoji: "📦",
-        colorClass: "bg-accent",
-        buttonClass: "btn-outline",
-        description: "Curated seasonal organic veggie boxes",
-    },
-];
 
 function CategoryCard({ c, idx }: { c: Card; idx: number }) {
     const { ref, visible } = useReveal<HTMLDivElement>();
@@ -141,22 +97,7 @@ function CategoryCard({ c, idx }: { c: Card; idx: number }) {
         </div>
     );
 }
+export {
+    CategoryCard
 
-export default function Category() {
-    return (
-        <section className="w-[92%] mx-auto mt-10">
-            <h2
-                className="text-3xl font-bold mb-6"
-                style={{ color: "hsl(var(--text-primary))" }}
-            >
-                Fresh Categories
-            </h2>
-
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-                {CARDS.map((c, i) => (
-                    <CategoryCard key={c.title} c={c} idx={i} />
-                ))}
-            </div>
-        </section>
-    );
-}
+}; // e:\ReacMain\Next js\vegetable\vegetable-shop\frontend\src\widgets\home-widgets\catagory\ui\catagory-card.tsx
