+++ frontend/src/widgets/home-widgets/customer-reviews/ui/customer-reviews.tsx
