+++ frontend/src/widgets/home-widgets/freshly-arrived/ui/freshly-arrived.tsx
