+++ frontend/src/widgets/home-widgets/index.ts
@@ -1,0 +1,5 @@
+export * from "./catagory/ui/catagry";
+export * from "./hero-slider/ui/Hero";
+export * from './seasonal-section/ui/seasonal-section'
+export * from "./customer-reviews/ui/customer-reviews"
+export * from "./freshly-arrived/ui/freshly-arrived"
