+++ frontend/src/widgets/home-widgets/seasonal-section/ui/seasonal-section.tsx
