+++ frontend/src/widgets/sidebar/ui/nav-item.tsx
