+++ frontend/src/widgets/sidebar/ui/sidebar.tsx
